Make cms dependency soft
