+++ djangocms_text/widgets.py
@@ -1,6 +1,7 @@
 import json
 import uuid
 from copy import deepcopy
+from functools import cache
 from itertools import groupby
 from typing import Union
 
@@ -9,6 +10,7 @@ from django.conf import settings
 from django.core.serializers.json import DjangoJSONEncoder
 from django.db import models
 from django.template.loader import render_to_string
+from django.urls.exceptions import NoReverseMatch
 from django.utils.safestring import mark_safe
 from django.utils.translation.trans_real import get_language, gettext
 
@@ -16,11 +18,25 @@ from cms.utils.urlutils import admin_reverse, static_with_version
 
 from . import settings as text_settings
 from .editors import DEFAULT_TOOLBAR_CMS, DEFAULT_TOOLBAR_HTMLField, get_editor_base_config, get_editor_config
-from .utils import cms_placeholder_add_plugin, get_url_endpoint
+from .utils import cms_placeholder_add_plugin
 
 
 rte_config = get_editor_config()
 #: The configuration for the text editor widget
+
+
+@cache
+def get_url_endpoint():
+    """Get the url for dynamic liks for cms plugins and HTMLFields"""
+    from django.contrib.admin import site
+
+    for model_admin in site._registry.values():
+        if hasattr(model_admin, "global_link_url_name"):
+            return admin_reverse(model_admin.global_link_url_name)
+    try:
+        return admin_reverse("djangocms_text_textplugin_get_available_urls")
+    except NoReverseMatch:
+        return None
 
 
 class TextEditorWidget(forms.Textarea):
